fix: 修正TeacherTest.java中status字段的数据类型
- 将status字段的数据类型从int改为short，以符合数据模型定义
- 修改涉及多个测试用例中的status字段设置

+++ src/test/java/com/frontleaves/scheduling/dao/TeacherTest.java
@@ -70,7 +70,7 @@ class TeacherTest {
                 .setPassword(PasswordUtil.encrypt("123456Aa"))
                 .setEmail("teacherDAOTest@qwer.com")
                 .setPhone("14452873800")
-                .setStatus(1)
+                .setStatus((short) 1)
                 .setBan(0)
                 .setRoleUuid(SystemConstant.getRoleTeacher())
                 .setPermission("[\"user:role:edit\"]");
@@ -192,7 +192,7 @@ class TeacherTest {
                 .setPassword(PasswordUtil.encrypt("123456Aa"))
                 .setEmail("teacherDAONewTestUser@qwer.com")
                 .setPhone("15859273800")
-                .setStatus(1)
+                .setStatus((short) 1)
                 .setBan(0)
                 .setRoleUuid(SystemConstant.getRoleTeacher())
                 .setPermission("[\"user:role:edit\"]");
@@ -221,7 +221,7 @@ class TeacherTest {
                 .setPassword(PasswordUtil.encrypt("123456Aa"))
                 .setEmail("logicUserTest@test.com")
                 .setPhone("13800000000")
-                .setStatus(1)
+                .setStatus((short) 1)
                 .setBan(0)
                 .setPermission("[\"user:unit:department:tag:category:delete\"]")
                 .setRoleUuid(SystemConstant.getRoleTeacher());
